Move CValue to separate module
